Reorganizing for saga
